requirements.txt
async-generator==1.10
attrs==22.2.0
black==23.1.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
exceptiongroup==1.1.0
h11==0.14.0
idna==3.4
mypy-extensions==1.0.0
outcome==1.2.0
packaging==23.0
pathspec==0.11.0
platformdirs==3.0.0
PySocks==1.7.1
python-dotenv==0.21.1
PyYAML==6.0
requests==2.28.2
selenium==4.13.0
sniffio==1.3.0
sortedcontainers==2.4.0
tomli==2.0.1
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.5.0
urllib3==1.26.14
webdriver-manager==4.0.1
wsproto==1.2.0
undetected-chromedriver==3.5.3
sendgrid~=6.11.0
APScheduler~=3.10.4